[Quality] fix warning about the memory management of Datapoint of Reading object

+++ include/iec61850_client_config.h
@@ -21,7 +21,7 @@
 
 struct ServerConnectionParameters {
     std::string ipAddress;
-    uint16_t    mmsPort;
+    int mmsPort;
 };
 
 struct ExchangedData {
